requirements.txt
langchain
langchain-openai
langchain-chroma
openai
chromadb
streamlit
python-dotenv
arxiv
